Create separate folder for queries
