+++ react/src/views/Distributions/DistroEventView/components/State2Packing/components/PackingAddBoxOrItemsForPackingListEntryOverlay/PackingScanBoxOrFindByLabelOverlayContent.tsx
@@ -13,7 +13,7 @@ import {
   GetBoxLabelIdentifierForQrCodeQuery,
   GetBoxLabelIdentifierForQrCodeQueryVariables
 } from "types/generated/graphql";
-import { BOX_DETAILS_BY_LABEL_IDENTIFIER_QUERY, GET_BOX_LABEL_IDENTIFIER_BY_QR_CODE } from "utils/queries";
+import { BOX_DETAILS_BY_LABEL_IDENTIFIER_QUERY, GET_BOX_LABEL_IDENTIFIER_BY_QR_CODE } from "queries/queries";
 import { BoxData, IPackingListEntry } from "views/Distributions/types";
 
 interface PackingScanBoxOrFindByLabelOverlayProps {
